+++ libraries/raw-rs/src/metadata/mod.rs
@@ -1,1 +1,3 @@
+pub mod camera_data;
 pub mod identify;
+
