+++ libraries/raw-rs/src/postprocessing/mod.rs
@@ -1,3 +1,4 @@
 pub mod convert_to_rgb;
 pub mod gamma_correction;
+pub mod record_histogram;
 pub mod transform;
