+++ libraries/raw-rs/src/preprocessing/mod.rs
@@ -1,4 +1,3 @@
-pub mod camera_data;
 pub mod scale_to_16bit;
 pub mod scale_white_balance;
 pub mod subtract_black;
